+++ src/main/java/seedu/toluist/controller/DeleteTaskController.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.logging.Logger;
 
 import seedu.toluist.commons.core.LogsCenter;
-import seedu.toluist.commons.util.StringUtil;
+import seedu.toluist.controller.commons.IndexTokenizer;
 import seedu.toluist.controller.commons.TaskTokenizer;
 import seedu.toluist.dispatcher.CommandResult;
 import seedu.toluist.model.Task;
@@ -39,7 +39,7 @@ public class DeleteTaskController extends Controller {
         HashMap<String, String> tokens = taskTokenizer.tokenize(command, true, false);
 
         String indexToken = tokens.get(TaskTokenizer.TASK_VIEW_INDEX);
-        List<Integer> indexes = StringUtil.splitIndexes(indexToken, todoList.getTasks().size());
+        List<Integer> indexes = IndexTokenizer.splitIndexes(indexToken, todoList.getTasks().size());
         List<Task> tasks = uiStore.getTasks(indexes);
         commandResult = delete(todoList, tasks);
